requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
pydantic-settings>=2.2,<3.0
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0,<4.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.11
